# tacocloud/tacocloud-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: TacoCloud course API
  version: 1.0.0
  description: 'TC-01 to TC-36. Money: MXN, sum of ingredient prices, HALF_UP at 2
    decimals. /api is a transitional alias with Deprecation: true and Link to this
    spec; no removal date is committed. Use /api/v1. Catalog metadata is academic
    and does not replace cross-contamination controls. Runtime Spring Boot 2.5 / MVC.'
servers:
- url: /api/v1
- url: /api
  description: Deprecated compatibility alias
paths:
...
  /orders:
    post:
      operationId: post_orders
      summary: Place order idempotently
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{8,100}$
        description: Scoped to authenticated user. Identical request replays its original
          response for at least 168 hours. Different payload returns 409. An aborted
          transaction leaves no IN_PROGRESS record.
      responses:
        '201':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
  /orders/quote:
    post:
      operationId: post_orders_quote
      summary: Quote without side effects
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
  /orders/fromEmail:
    post:
      operationId: post_orders_fromEmail
      summary: Place email order (ADMIN)
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '201':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailOrder'
  /orders/{id}:
    put:
      operationId: put_orders_id
      summary: put own order
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderReplacement'
    patch:
      operationId: patch_orders_id
      summary: patch own order
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPatch'
    delete:
      operationId: delete_orders_id
      summary: delete own order
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '204':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
  /orders/{id}/cancel:
    post:
      operationId: post_orders_id_cancel
      summary: Cancel own order before preparation
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
  /orders/{id}/status:
    patch:
      operationId: patch_orders_id_status
      summary: Change order status (ADMIN)
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transition'
  /orders/{id}/reorder:
    post:
      operationId: post_orders_id_reorder
      summary: Reorder with current rules
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{8,100}$
        description: Scoped to authenticated user. Identical request replays its original
          response for at least 168 hours. Different payload returns 409. An aborted
          transaction leaves no IN_PROGRESS record.
      responses:
        '201':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
...
  /users/me/orders/{id}:
    get:
      operationId: get_users_me_orders_id
      summary: Own order detail
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
...
  /kitchen/orders/{id}/status:
    patch:
      operationId: patch_kitchen_orders_id_status
      summary: Advance kitchen order (KITCHEN)
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KitchenOrder'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transition'
...
components:
  schemas:
...
    Status:
      type: string
      enum:
      - CREATED
      - ACCEPTED
      - PREPARING
      - READY
      - OUT_FOR_DELIVERY
      - DELIVERED
      - CANCELLED
...
    OrderCreateRequest:
      type: object
      properties:
        deliveryName:
          type: string
          minLength: 1
          maxLength: 100
        deliveryStreet:
          type: string
          minLength: 1
          maxLength: 200
        deliveryCity:
          type: string
          minLength: 1
          maxLength: 100
        deliveryState:
          type: string
          minLength: 1
          maxLength: 100
        deliveryZip:
          type: string
          pattern: ^[A-Za-z0-9 -]{3,12}$
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemRequest'
          minItems: 1
          maxItems: 50
        paymentMethodId:
          type: string
          minLength: 1
        couponCode:
          type: string
          maxLength: 40
          nullable: true
      additionalProperties: false
      required:
      - deliveryName
      - deliveryStreet
      - deliveryCity
      - deliveryState
      - deliveryZip
      - items
      - paymentMethodId
...
    OrderPatch:
      type: object
      properties:
        deliveryName:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryStreet:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
        deliveryCity:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryState:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryZip:
          type: string
          pattern: ^[A-Za-z0-9 -]{3,12}$
          nullable: true
      additionalProperties: false
...
    OrderResponse:
      type: object
      properties:
        deliveryName:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryStreet:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
        deliveryCity:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryState:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryZip:
          type: string
          pattern: ^[A-Za-z0-9 -]{3,12}$
          nullable: true
        id:
          type: string
          nullable: true
        placedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/Status'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        subtotal:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        discount:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        total:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        currency:
          type: string
          enum:
          - MXN
        version:
          type: integer
          nullable: true
        paymentBrand:
          type: string
          nullable: true
        paymentLast4:
          type: string
          pattern: ^[0-9]{4}$
          nullable: true
      additionalProperties: false
      required:
      - status
      - items
      - subtotal
      - discount
      - total
      - currency
...
    EmailOrder:
      type: object
      properties:
        email:
          type: string
          format: email
        tacos:
          type: array
          items:
            $ref: '#/components/schemas/EmailTaco'
          minItems: 1
          maxItems: 50
      additionalProperties: false
      required:
      - email
      - tacos
...
    ApiProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        correlationId:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
        previousTotal:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        quote:
          $ref: '#/components/schemas/OrderResponse'
      additionalProperties: false
      required:
      - type
      - title
      - status
      - detail
      - instance
      - code
      - correlationId
      - violations
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID